+++ src/main/java/com/baichen/rpc/consumer/ConsumerProperties.java
@@ -6,8 +6,10 @@ import lombok.Data;
 @Data
 public class ConsumerProperties {
     private Integer workerThreadNum = 4;
-    private Integer connectTimeoutMillis = 5000;
-    private Integer waitResponseTimeoutMillis = 5000;
+    private Integer connectTimeoutMs = 5000;
+    private Long waitResponseTimeoutMs = 3000L;
+    private Long totalTimeoutMs = 10000L;
     private String loadBalancePolicy = "roundRobin";
+    private String retryPolicy = "forkAll";
     private ServiceRegistryConfig serviceRegistryConfig;
 }