+++ src/main/java/com/baichen/rpc/retry/RetryPolicy.java
@@ -1,0 +1,8 @@
+package com.baichen.rpc.retry;
+
+import com.baichen.rpc.message.Response;
+
+public interface RetryPolicy {
+
+    Response retry(RetryContext context) throws Exception;
+}